Remove casts to long that cause > 4 GB files not to be correctly handled on Linux32, Win32 or Win64
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@26887 f0d54148-0727-0410-94bb-9a71ac55c965

+++ port/cpl_vsi_mem.cpp
@@ -541,7 +541,7 @@ int VSIMemFilesystemHandler::Stat( const char * pszFilename,
     }
     else
     {
-        pStatBuf->st_size = (long)poFile->nLength;
+        pStatBuf->st_size = poFile->nLength;
         pStatBuf->st_mode = S_IFREG;
     }
